コードの共通部分を関数としてまとめた
